recommendation/requirements.txt
flask
scikit-surprise
flask-cors
pandas
numpy
pymongo